+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ### Fixed
 - Fixed element events ignoring `criticalEnabled` setting. [#5]
 - Critical CSS is on longer output when critical is being generated.
+- Fixed handling of pseudo elements (pseudo classes still an issue ☹️) [#3]
 
+[#3]: https://github.com/ethercreative/critical/issues/3
 [#5]: https://github.com/ethercreative/critical/issues/5
 
 ## 0.0.1-alpha.2 - 2018-09-19